backend/requirements.txt
flask
flask-mysqldb
werkzeug
PyJWT
bcrypt
reportlab
